+++ src/main/java/by/myProject/model/service/RoleService.java
@@ -8,10 +8,10 @@ public interface RoleService {
 
     List<Role> findAllRoles();
     Role findByType(String typeRole);
-    Role findById(int id);
+    Role findById(Long id);
     void update(Role role);
     void save (Role role);
-    void deleteById(int id);
+    void deleteById(Long id);
